Split FunctionObject into two structs
